+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 # Mockitools Changelog
 
 ## [Unreleased]
+### Added
+- [#43](https://github.com/picimako/mockitools/issues/43): Add support for converting the `mockMaker` mock setting between the `@Mock` annotation and `MockSettings`.
+- Added missing support for the conversion of the `strictness` mock setting from `@Mock` to `MockSettings`.
+
 ### Changed
 - Dropped support for IJ-2021.2.
 - Added support for IJ-2022.3 EAPs.